Edit profile center manager collector customer

+++ app/helpers/ImageUpload_Helper.php
@@ -21,11 +21,30 @@ function uploadImage($img, $img_name, $location) {
 }
 
 
-function updateImage($old,$img,$img_name,$location){
-    unlink($old);
-    $target=PBROOT.$location.$img_name;
-    return move_uploaded_file($img,$target);
+function updateImage($old, $img, $img_name, $location) {
+    
+    if (!file_exists($img)) {
+        return false;
+    }
+
+    $allowed_types = ['image/png', 'image/jpeg'];
+    $file_info = finfo_open(FILEINFO_MIME_TYPE);
+    $mime_type = finfo_file($file_info, $img);
+    finfo_close($file_info);
+
+    if (!$mime_type || !in_array($mime_type, $allowed_types)) {
+        return false;
+        die();
+    }
+    if (file_exists($old)) {
+        unlink($old);
+    }
+
+    $target = PBROOT . $location . $img_name;
+
+    return move_uploaded_file($img, $target);
 }
+
 
 function deleteImage($img){
     if(unlink($img)){
